+++ src/ducks/Studio/Template/DialogLoadTemplate/Template.module.scss
@@ -1,11 +1,6 @@
 @import '~@santiment-network/ui/mixins';
 
 .wrapper {
-  width: 515px;
-  padding: 16px 12px;
-}
-
-.template {
   padding: 12px 16px;
   border-radius: 4px;
   user-select: none;
@@ -47,6 +42,10 @@
   width: 20px;
 }
 
+.left {
+  pointer-events: none;
+}
+
 .menu {
   width: 32px;
   margin-left: 6px;
@@ -61,10 +60,4 @@
 
 .toggle {
   margin-left: 16px;
-}
-
-.search {
-  padding: 8px 11px;
-  background: var(--athens);
-  border-bottom: 1px solid var(--porcelain);
 }
