+++ src/ducks/Studio/Template/DialogLoadTemplate/index.module.scss
@@ -1,0 +1,10 @@
+.wrapper {
+  width: 515px;
+  padding: 16px 12px;
+}
+
+.search {
+  padding: 8px 11px;
+  background: var(--athens);
+  border-bottom: 1px solid var(--porcelain);
+}
